fix: enum messageError and messageSystem + decorator ApiData
