+++ src/modules/chats/controllers/chats.controller.ts
@@ -1,50 +1,10 @@
-import { applyDecorators, Body, Controller, Get, Headers, Param, Post, Query, Type } from '@nestjs/common';
-import { ApiExtraModels, ApiOkResponse, getSchemaPath } from '@nestjs/swagger';
+import { Body, Controller, Get, Headers, Param, Post, Query } from '@nestjs/common';
 import { ChatsService } from '../services/chats.service';
 import { CreateOpenChatDto } from '../dto/requests/create-open-chat.dto';
 import { QueryGetChatsDto } from '../dto/requests/query-get-chats.dto';
 import { ChatEntity } from '../entities/chat.entity';
 import { DataResponse } from '../../../common/swagger/data-response.dto';
-import { MessageEntity } from '../entities/message.entity';
-
-export function ApiData(type: Type, isArray = false) {
-    const array = {
-        type: 'array',
-        items: { $ref: getSchemaPath(type) },
-    };
-
-    const notArray = { $ref: getSchemaPath(type) };
-
-    return applyDecorators(
-        ApiExtraModels(ChatEntity, MessageEntity),
-        ApiOkResponse({
-            schema: {
-                anyOf: [
-                    {
-                        type: 'object',
-                        properties: {
-                            success: { type: 'boolean', example: true },
-                            data: isArray ? array : notArray,
-                        },
-                        required: ['success', 'data'],
-                    },
-
-                    {
-                        type: 'object',
-
-                        properties: {
-                            success: { type: 'boolean', example: false },
-                            data: {
-                                type: 'string',
-                            },
-                        },
-                        required: ['success', 'data'],
-                    },
-                ],
-            },
-        }),
-    );
-}
+import { ApiData } from '../../../common/swagger/ decorator';
 
 @Controller('chats')
 export class ChatsController {
