+++ src/modules/chats/types/message-error-language.enum.ts
@@ -1,5 +1,5 @@
 export enum MessageErrorLanguageEnum {
-    CHAT_WITH_ID_NOT_FOUND = `Чат с данным id не найден`,
-    CHAT_NOT_FOUND = 'Чат не найден',
-    PARENTAL_MESSAGE_NOT_FOUND = 'Родительское сообщение не найдено',
+    CHAT_WITH_ID_NOT_FOUND = `chat_with_id_not_found`,
+    CHAT_NOT_FOUND = 'chat_nor_found',
+    PARENTAL_MESSAGE_NOT_FOUND = 'Parent_message_not_found',
 }
